requirements.txt
line-bot-sdk==3.11.0
flask==3.0.0
anthropic
openai